Error fix
fixed the error of 'can't run reduce method on null value' when opening
up the app on a new computer. Basically the same as the bill totalSum
method.
if there are no bills and no income, then the reduce method can't be
called. So I just added a conditional to check to see if values exist.
If so, run the code, if not, do nothing. Fixed